backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Altron Backend Service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "fastapi>=0.115.12",
    "pydantic-settings>=2.9.1",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pyright>=1.1.400",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "pytest-xdist>=3.7.0",
    "requests-mock>=1.12.1",
    "ruff>=0.11.10",
]

# Pytest settings
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing -n auto"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "raise ImportError",
    "except ImportError:",
    "if TYPE_CHECKING:",
]

# Pyright settings
[tool.pyright]
include = ["src", "tests"]
exclude = [
    "**/__pycache__",
    "**/.mypy_cache",
    "**/.pytest_cache",
    "**/.ruff_cache",
]

# Strict type checking settings
typeCheckingMode = "strict"
reportMissingImports = true
reportMissingTypeStubs = true
reportUnknownParameterType = true
reportUnknownArgumentType = true
reportUnknownLambdaType = true
reportUnknownVariableType = true
reportUnknownMemberType = true
reportMissingTypeArgument = true
reportInvalidTypeVarUse = true
reportUnnecessaryIsInstance = true
reportUnnecessaryCast = true
reportUnnecessaryComparison = true
reportUnnecessaryContains = true
reportUnusedExpression = true
reportWildcardImportFromLibrary = true
reportOptionalSubscript = true
reportOptionalMemberAccess = true
reportOptionalCall = true
reportOptionalIterable = true
reportOptionalContextManager = true
reportOptionalOperand = true
strictListInference = true
strictDictionaryInference = true
strictSetInference = true
strictParameterNoneValue = true
enableTypeIgnoreComments = true

# Python version and environment settings
pythonVersion = "3.12"
pythonPlatform = "Linux"
